ATM math problem
User inputs two values (an account balance, withdraw value) then the system checks to see if the withdraw is a multiple of 5 and if the balance on the account remaining after a withdrawal plus the transaction fee of (.50) is greater than zero. If true the transaction is performed, if either option is false than the transaction is not performed.